Update es_ES.json
To complete Cesar's work on Spanish translation. I havent found a suitable translation for token. I think we should keep it that way or we may risk generating some confusion to Spanish speaking users. I'm not sure of Power translation because I do not know the UI context. Please check it.